# webapp/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: Mojito API Documentation
  title: Mojito API
  version: 0.111-SNAPSHOT
servers:
- description: Generated server url
  url: http://localhost:8080
paths:
...
  /api/deltas/date:
    get:
      operationId: getDeltasFromDate
      parameters:
      - in: query
        name: repositoryId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: bcp47Tags
        required: false
        schema:
          type: array
          items:
            type: string
      - in: query
        name: fromDate
        required: false
        schema:
          type: string
          format: date-time
      - in: query
        name: toDate
        required: false
        schema:
          type: string
          format: date-time
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          default:
          - "id,DESC"
          items:
            type: string
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/PageTextUnitVariantDeltaDTO"
          description: OK
      summary: Get paginated Text Unit Variants Deltas for a given set of parameters
      tags:
      - delta-ws
  /api/deltas/state:
    get:
      operationId: getDeltasForRuns
      parameters:
      - in: query
        name: repositoryId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: pushRunIds
        required: true
        schema:
          type: array
          items:
            type: integer
            format: int64
      - in: query
        name: bcp47Tags
        required: false
        schema:
          type: array
          items:
            type: string
      - in: query
        name: pullRunIds
        required: false
        schema:
          type: array
          items:
            type: integer
            format: int64
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/DeltaResponseDTO"
          description: OK
      summary: Get Delta Content for a given set of parameters
      tags:
      - delta-ws
  /api/drops:
    get:
      operationId: getDrops
      parameters:
      - in: query
        name: repositoryId
        required: false
        schema:
          type: integer
          format: int64
      - in: query
        name: imported
        required: false
        schema:
          type: boolean
      - in: query
        name: canceled
        required: false
        schema:
          type: boolean
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          default:
          - "id,DESC"
          items:
            type: string
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/PageDrop_DropSummary"
          description: OK
      tags:
      - drop-ws
  /api/drops/cancel:
    post:
      operationId: cancelDrop
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelDropConfig"
        required: true
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/CancelDropConfig"
          description: OK
      tags:
      - drop-ws
  /api/drops/complete/{dropId}:
    post:
      operationId: completeDropById
      parameters:
      - in: path
        name: dropId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      tags:
      - drop-ws
  /api/drops/export:
    post:
      operationId: exportDrop
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportDropConfig"
        required: true
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/ExportDropConfig"
          description: OK
      tags:
      - drop-ws
  /api/drops/import:
    post:
      operationId: importDrop
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportDropConfig"
        required: true
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/ImportDropConfig"
          description: OK
      tags:
      - drop-ws
  /api/drops/importXliff:
    post:
      operationId: importXliff
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportXliffBody"
        required: true
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/ImportXliffBody"
          description: OK
      tags:
      - drop-ws
...
components:
  schemas:
...
    Authority:
      type: object
      properties:
        authority:
          type: string
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        lastModifiedDate:
          type: integer
          format: int64
          example: 1715699917000
...
    CancelDropConfig:
      type: object
      properties:
        dropId:
          type: integer
          format: int64
        pollableTask:
          $ref: "#/components/schemas/PollableTask"
      required:
      - dropId
...
    DeltaLocaleDataDTO:
      type: object
      properties:
        translationsByTextUnitName:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/DeltaTranslationDTO"
    DeltaMetadataDTO:
      type: object
      properties:
        fromDate:
          type: integer
          format: int64
          example: 1715699917000
        toDate:
          type: integer
          format: int64
          example: 1715699917000
    DeltaResponseDTO:
      type: object
      properties:
        content:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/DeltaLocaleDataDTO"
        metadata:
          $ref: "#/components/schemas/DeltaMetadataDTO"
    DeltaTranslationDTO:
      type: object
      properties:
        deltaType:
          type: string
          enum:
          - UNKNOWN
          - NEW_TRANSLATION
          - UPDATED_TRANSLATION
        text:
          type: string
    Drop_DropSummary:
      type: object
      properties:
        canceled:
          type: boolean
        createdByUser:
          $ref: "#/components/schemas/User_DropSummary"
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        dropExporterConfig:
          type: string
        exportFailed:
          type: boolean
        exportPollableTask:
          $ref: "#/components/schemas/PollableTask_DropSummary"
        id:
          type: integer
          format: int64
        importFailed:
          type: boolean
        importPollableTask:
          $ref: "#/components/schemas/PollableTask_DropSummary"
        lastImportedDate:
          type: integer
          format: int64
          example: 1715699917000
        name:
          type: string
        partiallyImported:
          type: boolean
        repository:
          $ref: "#/components/schemas/Repository_DropSummary"
        translationKits:
          type: array
          items:
            $ref: "#/components/schemas/TranslationKit_DropSummary"
          uniqueItems: true
    ExportDropConfig:
      type: object
      properties:
        dropId:
          type: integer
          format: int64
        locales:
          type: array
          items:
            type: string
        pollableTask:
          $ref: "#/components/schemas/PollableTask"
        repositoryId:
          type: integer
          format: int64
        type:
          type: string
          enum:
          - TRANSLATION
          - REVIEW
        uploadTime:
          type: integer
          format: int64
          example: 1715699917000
        useInheritance:
          type: boolean
      required:
      - repositoryId
...
    ImportDropConfig:
      type: object
      properties:
        dropId:
          type: integer
          format: int64
        pollableTask:
          $ref: "#/components/schemas/PollableTask"
        repositoryId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - TRANSLATION_NEEDED
          - REVIEW_NEEDED
          - MT_TRANSLATED
          - MT_REVIEW_NEEDED
          - APPROVED
          - OVERRIDDEN
      required:
      - dropId
      - repositoryId
...
    ImportXliffBody:
      type: object
      properties:
        importStatus:
          type: string
          enum:
          - TRANSLATION_NEEDED
          - REVIEW_NEEDED
          - MT_TRANSLATED
          - MT_REVIEW_NEEDED
          - APPROVED
          - OVERRIDDEN
        repositoryId:
          type: integer
          format: int64
        translationKit:
          type: boolean
        xliffContent:
          type: string
...
    Locale_DropSummary:
      type: object
      properties:
        bcp47Tag:
          type: string
        id:
          type: integer
          format: int64
...
    PageDrop_DropSummary:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Drop_DropSummary"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject_DropSummary"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageTextUnitVariantDeltaDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/TextUnitVariantDeltaDTO"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
...
    PageableObject_DropSummary:
      type: object
...
    PollableTask:
      type: object
      properties:
        allFinished:
          type: boolean
        createdByUser:
          $ref: "#/components/schemas/User"
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        errorMessage:
          type: string
        errorStack:
          type: string
        expectedSubTaskNumber:
          type: integer
          format: int32
        finishedDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        lastModifiedDate:
          type: integer
          format: int64
          example: 1715699917000
        message:
          type: string
        name:
          type: string
        subTasks:
          type: array
          items:
            $ref: "#/components/schemas/PollableTask"
          uniqueItems: true
    PollableTask_DropSummary:
      type: object
      properties:
        allFinished:
          type: boolean
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
...
    Repository_DropSummary:
      type: object
      properties:
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        name:
          type: string
...
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    TextUnitVariantDeltaDTO:
      type: object
      properties:
        bcp47Tag:
          type: string
        content:
          type: string
        deltaType:
          type: string
          enum:
          - UNKNOWN
          - NEW_TRANSLATION
          - UPDATED_TRANSLATION
        textUnitName:
          type: string
...
    TranslationKit_DropSummary:
      type: object
      properties:
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        imported:
          type: boolean
        locale:
          $ref: "#/components/schemas/Locale_DropSummary"
        type:
          type: string
          enum:
          - TRANSLATION
          - REVIEW
        wordCount:
          type: integer
          format: int64
...
    User:
      type: object
      properties:
        authorities:
          type: array
          items:
            $ref: "#/components/schemas/Authority"
          uniqueItems: true
        commonName:
          type: string
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        enabled:
          type: boolean
        givenName:
          type: string
        id:
          type: integer
          format: int64
        lastModifiedDate:
          type: integer
          format: int64
          example: 1715699917000
        partiallyCreated:
          type: boolean
        password:
          type: string
          writeOnly: true
        surname:
          type: string
        username:
          type: string
...
    User_DropSummary:
      type: object
      properties:
        commonName:
          type: string
        createdDate:
          type: integer
          format: int64
          example: 1715699917000
        id:
          type: integer
          format: int64
        username:
          type: string
...
  securitySchemes:
    cookieAuth:
      in: cookie
      name: JSESSIONID
      type: apiKey